FIX116 - Rename the loaders folder to core
