+++ readme.md
@@ -30,6 +30,7 @@
 
 ## Folder structure
 
+    |-- data/                     # Database files for the mongo docker container
     |-- public/                   # Static assets & symlinks for locally stored files
     |-- src/
     |   |-- @types/               # Configurations for declaration merging
@@ -38,7 +39,7 @@
     |   |   |-- middlewares/      # Interceptor filters before router functions
     |   |   |-- routes/           # Routes for the REST API
     |   |-- configs/              # Configurations retrieved from env variables
-    |   |-- loaders/              # Modules split for easier startup of application
+    |   |-- core/                 # Core modules split for easier startup of the application
     |   |-- models/               # Mongoose models as data access layer
     |   |-- services/             # Main business logic code as service layer
     |   |-- shared/               
@@ -50,7 +51,10 @@
     |   |-- utils/                # Local utility functions for application
     |   |-- app.ts                # Application main file
     |-- storage/                  # Local storage for cache, files and logs
-    |-- test/                     # Specs for unit tests
+    |-- tests/                    # Specs for unit tests
+    |-- tools/                    
+    |   |-- docker/               # Dockerfiles for custom images
+    |   |-- nginx/                # Configuration files for the nginx reverse proxy
     |-- views/                    # Templates for emails & pdf files(e.g. invoices)
     |-- .env.example              # Local environment config sample
