+++ src/core/aws-sthree.ts
